feat: fix storybook dark mode + improve a few stories
