+++ package.json
@@ -26,10 +26,6 @@
   },
   "dependencies": {
     "@circle-fin/developer-controlled-wallets": "^7.0.0",
-    "@radix-ui/react-dialog": "^1.1.2",
-    "@radix-ui/react-label": "^2.1.0",
-    "@radix-ui/react-select": "^2.1.2",
-    "@radix-ui/react-slot": "^1.1.0",
     "@web3icons/react": "^3.16.0",
     "lucide-react": "^0.462.0",
     "react": "^18.2.0",
