+++ packages/circle-react-elements/src/components/TokenSelect/TokenSelect.stories.ts
@@ -52,9 +52,6 @@ const exampleBalances: Balance[] = [
 const meta = {
   title: 'Token Components/TokenSelect',
   component: TokenSelect,
-  parameters: {
-    layout: 'centered', // Centers the component in the story canvas
-  },
   tags: ['autodocs'],
 } satisfies Meta<typeof TokenSelect>;
